refactor(hub-client): move actor palette helpers into utils/
`actorColor` and `fnv1aHex8` are pure utilities with a strict
Rust-sibling drift contract (`crates/quarto-core/src/attribution/
palette.rs`). They previously lived inside the React-bearing
`hooks/useReplayMode.ts`, which forced `useAttribution.ts` to import
them from a sibling hook module. Move both functions and their
drift-mitigation tests to `utils/palette.ts` + `utils/palette.test.ts`
so the two real call sites (`useReplayMode` / `useAttribution`) and
`ReplayDrawer.tsx` import from a single non-React module.

No behavior change; both helpers are byte-for-byte identical and the
test vectors are unchanged.
