+++ hub-client/src/utils/palette.ts
@@ -1,0 +1,58 @@
+/**
+ * Deterministic actor-colour helpers shared between the replay drawer
+ * (`useReplayMode` / `ReplayDrawer`) and the attribution producer
+ * (`useAttribution`). Both must produce identical visual output for the
+ * same actor — colours seen during replay must match colours seen on
+ * Authorship overlays.
+ *
+ * **Drift discipline.** Both functions MUST stay bit-for-bit identical
+ * with their Rust siblings in
+ * `crates/quarto-core/src/attribution/palette.rs`. The Rust side
+ * colours git-blame author emails (via `--attribution=git`); this TS
+ * side colours Automerge actor IDs. They share `actor_color` /
+ * `actorColor` and `fnv1a_hex8` / `fnv1aHex8` definitions; the test
+ * suite in `palette.test.ts` pins reference vectors that match
+ * `palette.rs::tests`.
+ */
+
+/**
+ * Deterministic colour from an actor hash string.
+ *
+ * Formula: parse the first 6 hex chars of the actor ID as an integer,
+ * mod 360, emit `hsl(<hue>, 60%, 55%)`. Non-hex input (or an empty
+ * string) collapses to hue `0`.
+ */
+export function actorColor(actor: string): string {
+  const hue = parseInt(actor.slice(0, 6), 16) % 360;
+  return `hsl(${hue}, 60%, 55%)`;
+}
+
+/**
+ * 32-bit FNV-1a hash, formatted as a left-padded 8-char hex string.
+ *
+ * Used to reduce an arbitrary actor string (e.g. an Automerge actor
+ * ID whose first 6 chars aren't guaranteed hex) to a hex-prefix-safe
+ * input for `actorColor`. Caller: the attribution producer in
+ * `useAttribution`, when synthesising the `(name, color)` fallback
+ * identity for actors with no profile metadata.
+ *
+ * Not cryptographic; deterministic and well-distributed for colour
+ * purposes.
+ */
+export function fnv1aHex8(s: string): string {
+  // Hash UTF-8 bytes to match Rust's `s.bytes()` iteration. JS strings
+  // are UTF-16 internally; using `charCodeAt` directly would diverge
+  // from Rust on any non-ASCII character.
+  const bytes = TEXT_ENCODER.encode(s);
+  let hash = 0x811c9dc5;
+  for (let i = 0; i < bytes.length; i++) {
+    hash ^= bytes[i];
+    // Multiply by 0x01000193 (16777619), masked back into a u32 each
+    // iteration. `Math.imul` does 32-bit multiplication; `>>> 0`
+    // coerces to unsigned.
+    hash = Math.imul(hash, 0x01000193) >>> 0;
+  }
+  return hash.toString(16).padStart(8, '0');
+}
+
+const TEXT_ENCODER = new TextEncoder();
